% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_choice.R
\name{approx_log_ml}
\alias{approx_log_ml}
\title{approx log marginal likelihood for the last fit in the list}
\usage{
approx_log_ml(fits)
}
\description{
approx log marginal likelihood for the last fit in the list
}
