% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_choice.R
\name{approx_log_ml_contrib}
\alias{approx_log_ml_contrib}
\title{approximate the log marginal likelihood
by using a Laplace approximation
approximates - log phi(0, 0, Sigma)}
\usage{
approx_log_ml_contrib(Sigma_inv)
}
\description{
approximate the log marginal likelihood
by using a Laplace approximation
approximates - log phi(0, 0, Sigma)
}
