% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{evaluate_with_extrap}
\alias{evaluate_with_extrap}
\title{if x outside knots, extrapolate constant at nearest knot point, rather than NA
(which is the default in orthogonalsplinebasis:::evaluate.SplineBasis}
\usage{
evaluate_with_extrap(object, x, deriv)
}
\description{
if x outside knots, extrapolate constant at nearest knot point, rather than NA
(which is the default in orthogonalsplinebasis:::evaluate.SplineBasis
}
