% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{find_pen_deviance_k}
\alias{find_pen_deviance_k}
\title{using deviance = -2 * l_pen so can minimize rather than maximise function
(so can use nlm for optimization)}
\usage{
find_pen_deviance_k(alpha_k, sp, X_k, S_k, fit_km1)
}
\description{
using deviance = -2 * l_pen so can minimize rather than maximise function
(so can use nlm for optimization)
}
