% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{find_samples}
\alias{find_samples}
\title{Find parametric bootstrap samples from a fitted model}
\usage{
find_samples(mod, n_samples)
}
\arguments{
\item{mod}{The fitted model, output from fit_flexl.}

\item{n_samples}{The number of samples to generate.}
}
\description{
Find parametric bootstrap samples from a fitted model
}
