% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_given_sp.R
\name{fit_0}
\alias{fit_0}
\title{Fit the mean-only mode, and set up for use in future fits
r is rank of S}
\usage{
fit_0(data, sp, basis)
}
\description{
Fit the mean-only mode, and set up for use in future fits
r is rank of S
}
