% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_flexl}
\alias{fit_flexl}
\title{start with first sp and k = 0
increase k until optimise log ML, for smallest sp
increase sp to next value
then check if we can decrease k. Continue decreasing k until
we optimize log ML
check if we have found a better log_ML at this value of sp
than we had found at previous value}
\usage{
fit_flexl(data, nbasis = 10, kmax = 10, lsp_poss = -5:15)
}
\description{
start with first sp and k = 0
increase k until optimise log ML, for smallest sp
increase sp to next value
then check if we can decrease k. Continue decreasing k until
we optimize log ML
check if we have found a better log_ML at this value of sp
than we had found at previous value
}
