% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_given_sp.R
\name{fit_given_fit_kp1}
\alias{fit_given_fit_kp1}
\title{fit model with k eigenfunctions, given model fit with k + 1 eigenfunctions, same sp}
\usage{
fit_given_fit_kp1(data, sp, k, fit_kp1, basis)
}
\description{
fit model with k eigenfunctions, given model fit with k + 1 eigenfunctions, same sp
}
