% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_given_sp.R
\name{fit_given_fit_other_sp}
\alias{fit_given_fit_other_sp}
\title{fit model, given fit_other_sp with same k but different sp
using hessian from fit_other_sp, do steps of Newton-Raphson with hessian H}
\usage{
fit_given_fit_other_sp(data, sp, fit_other_sp, basis)
}
\description{
fit model, given fit_other_sp with same k but different sp
using hessian from fit_other_sp, do steps of Newton-Raphson with hessian H
}
