% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted_flexl}
\alias{fitted_flexl}
\title{Fitted values from flexible models for longitudinal data}
\usage{
fitted_flexl(mod)
}
\arguments{
\item{mod}{The fitted model, output from fit_flexl}
}
\description{
Fitted values from flexible models for longitudinal data
}
