% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_covariance_update.R
\name{ldmvnorm}
\alias{ldmvnorm}
\title{Find log phi(z; 0, Sigma_k),
where z is fixed and stored
and
Sigma_k = Sigma_km1 + a a^T
for each new value of a.
Here info_km1 contains values needed corresponding to Sigma_km1 and z}
\usage{
ldmvnorm(a, info_km1)
}
\description{
Find log phi(z; 0, Sigma_k),
where z is fixed and stored
and
Sigma_k = Sigma_km1 + a a^T
for each new value of a.
Here info_km1 contains values needed corresponding to Sigma_km1 and z
}
