% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_covariance_update.R
\name{log_det_Sigma_grad}
\alias{log_det_Sigma_grad}
\title{Derivative of log(|Sigma_k|) with respect to a
(computed above as const + log c)}
\usage{
log_det_Sigma_grad(c, dc)
}
\description{
Derivative of log(|Sigma_k|) with respect to a
(computed above as const + log c)
}
