% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_y_given_sample}
\alias{predict_y_given_sample}
\title{Prediction for flexible models for longitudinal data, given pre-existing sample}
\usage{
predict_y_given_sample(sample, mod, newdata, deriv = FALSE)
}
\arguments{
\item{sample}{A parametric bootstrap sample, generated by find_samples}

\item{mod}{The fitted model, output from fit_flexl}

\item{newdata}{A data frame, containing the new
values of x, y and c at which to make predictions.}

\item{deriv}{If TRUE, output predictions for derivative
of subject-specific mean function}
}
\description{
Prediction for flexible models for longitudinal data, given pre-existing sample
}
