% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_covariance_update.R
\name{quad_form_grad}
\alias{quad_form_grad}
\title{Derivative of z^T (Sigma_k)^{-1} z with respect to a
(computed above as const - d^2 / c)}
\usage{
quad_form_grad(c, dc, d, dd)
}
\description{
Derivative of z^T (Sigma_k)^{-1} z with respect to a
(computed above as const - d^2 / c)
}
