% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{simulate_1dv}
\alias{simulate_1dv}
\title{Simulate a one-dimensional variation example}
\usage{
simulate_1dv(seed, beta0, beta1, sigma_u, sigma, n_clusters, n_obs_per_cluster)
}
\arguments{
\item{seed}{The random seed.}

\item{beta0}{The value of beta_0 in the model.}

\item{beta1}{The value of beta_1 in the model.}

\item{sigma_u}{The standard deviation of the random effects.}

\item{sigma}{The standard deviation of the observation error.}

\item{n_clusters}{The number of clusters or individuals.}

\item{n_obs_per_cluster}{The number of observations on each cluster or individual.}
}
\description{
Simulate a one-dimensional variation example
}
