% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_flexl}
\alias{simulate_flexl}
\title{simulate normalised data, using original x and c
output mu as "truth" for simulation studies}
\usage{
simulate_flexl(mod)
}
\description{
simulate normalised data, using original x and c
output mu as "truth" for simulation studies
}
