% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{simulate_ri}
\alias{simulate_ri}
\title{Simulate from a random intercept model}
\usage{
simulate_ri(seed, beta0, beta1, sigma_u, sigma, n_clusters, n_obs_per_cluster)
}
\arguments{
\item{seed}{The random seed.}

\item{beta0}{The population-level intercept.}

\item{beta1}{The population-level slope.}

\item{sigma_u}{The standard deviation of the random effects.}

\item{sigma}{The standard deviation of the observation error.}

\item{n_clusters}{The number of clusters or individuals.}

\item{n_obs_per_cluster}{The number of observations on each cluster or individual.}
}
\description{
Simulate from a random intercept model
}
