% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_flexl}
\alias{fit_flexl}
\title{Fit a flexible model for simple longitudinal data.}
\usage{
fit_flexl(data, nbasis = 10, kmax = 10, lsp_poss = -5:15, trace = FALSE)
}
\arguments{
\item{data}{A data frame, with columns c (identifying
the individual subjects), x (the time) and y (the response).}

\item{nbasis}{The number of spline basis functions.}

\item{kmax}{The maximum number of functional principal components
to allow.}

\item{lsp_poss}{The grid of possible values to consider for
log(gamma), the log of the smoothing parameter.}

\item{trace}{If TRUE, print out extra information.}
}
\value{
The fitted model.
}
\description{
Fit a flexible model for simple longitudinal data.
}
\examples{
data_full <- simulate_1dv(1, -0.5, 0.1, 0.5, 0.1, 20, 10)
data <- data_full$data
mod <- fit_flexl(data)
}
